Added maximum order limits for restaurants
added to the restaurant dashboard a text input to change the maximum daily order amount. this amount changes on the database linked to the restaurant.

moved the styling to a stylesheet at bottom for better readability

fixed the alerts to show on web

still to implement - restricting orders/marking restaurant as busy when this maximum has been met?