Working on restaurant daily limits
Have implemented code for completing the order which adds one to each restaurant contained in the orders daily order count. When the daily order count reaches the daily limit, the restaurant is marked as busy.

Bug: even though the restaurant is marked as busy on supabase, this is not reflected on the app?

Still to implement: resetting the daily orders and marking the restaurant as not busy for a new day?